#129 Refactor Redux / cleanup (#145)
* Create new userReducer

* Create new contactReducer

* Change all mapStateToProps functions to new reducers

* Externalize appActions & change imports
