+++ src/actions/userActions.test.js
@@ -6,20 +6,12 @@ import fetchMock from 'fetch-mock';
 import User from 'ownuser';
 import {
     SET_WEBID,
-    SET_CURRENT_CONTACT,
-    SET_SELECTION,
     FETCH_USER,
     FETCH_USER_SUCCESS,
     FETCH_USER_FAIL,
 } from './types';
-import {
-    setWebId,
-    setCurrentContact,
-    setSelection,
-    fetchUser,
-} from './UserActions';
+import { fetchUser, setWebId } from './userActions';
 
-import { user } from '../test/fixtures/user';
 import { card } from '../test/fixtures/card';
 
 const middlewares = [thunk];
@@ -53,29 +45,6 @@ describe('redux actions', () => {
                     payload: webId,
                 };
                 expect(setWebId(webId)).toEqual(expectedAction);
-            });
-        });
-        describe('setCurrentContact', () => {
-            it('should create action to set currentContact', () => {
-                const expectedAction = {
-                    type: SET_CURRENT_CONTACT,
-                    payload: user,
-                };
-                expect(setCurrentContact(user)).toEqual(expectedAction);
-            });
-        });
-        describe('setSelection', () => {
-            it('should create action to set selection', () => {
-                const selection = [
-                    'https://bejow.owntech.de/favicon.ico',
-                    'https://bejow.owntech.de/profile/',
-                    'https://bejow.owntech.de/Buttons.svg',
-                ];
-                const expectedAction = {
-                    type: SET_SELECTION,
-                    payload: selection,
-                };
-                expect(setSelection(selection)).toEqual(expectedAction);
             });
         });
     });
