+++ src/functional_components/ContactListItem/ContactListItem.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 import styles from './ContactListItem.module.css';
 import defaultIcon from '../../assets/icons/defaultUserPic2.png';
 import Delete from '../../assets/svgIcons/Delete';
-import { removeContact } from '../../actions/UserActions';
+import { removeContact } from '../../actions/contactActions';
 import { withRouter } from 'react-router-dom';
 
 const ContactListItem = ({
@@ -54,7 +54,7 @@ ContactListItem.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-    webId: state.app.webId,
+    webId: state.user.webId,
 });
 
 export default connect(
