+++ src/reducers/appReducer.test.js
@@ -1,4 +1,4 @@
-import { isContact } from './AppReducer';
+import { isContact } from './contactReducer';
 import { contacts } from '../test/fixtures/contacts';
 describe('AppReducer', () => {
     describe('isContact() Selector', () => {
