+++ src/reducers/index.js
@@ -1,5 +1,10 @@
 import { combineReducers } from 'redux';
-import appReducer from './AppReducer';
+import appReducer from './appReducer';
+import userReducer from './userReducer';
+import contactReducer from './contactReducer';
+
 export default combineReducers({
     app: appReducer,
+    user: userReducer,
+    contact: contactReducer,
 });
